Fix warning for unsigned comparision in gcc9 for rest_client [753] (#1387)

+++ tiledb/sm/rest/rest_client.cc
@@ -434,7 +434,7 @@ size_t RestClient::post_data_write_cb(
 
   bytes_processed += length;
 
-  assert(bytes_processed == content_nbytes);
+  assert(static_cast<size_t>(bytes_processed) == content_nbytes);
   return std::max(bytes_processed, 0L);
 }
